+++ sources/BaseDestination.swift
@@ -90,6 +90,8 @@ public class BaseDestination: Hashable, Equatable {
     public var detailOutput = true
     /// adds colored log levels where possible
     public var colored = true
+    /// colors entire log
+    public var coloredLines = false
     /// runs in own serial background thread for better performance
     @available(*, deprecated=0.0, message="use the var executionContext instead!")
     public var asynchronously: Bool {
@@ -147,11 +149,11 @@ public class BaseDestination: Hashable, Equatable {
     // For a colored log level word in a logged line
     // XCode RGB colors
     public struct LevelColor {
-        public var Verbose = "fg200,200,200;"     // silver
-        public var Debug = "fg0,255,0;"           // green
-        public var Info = "fg0,0,255;"           // blue
-        public var Warning = "fg255,255,0;"       // yellow
-        public var Error = "fg255,0,0;"           // red
+        public var Verbose = "fg150,178,193;"     // silver
+        public var Debug = "fg32,155,124;"        // green
+        public var Info = "fg70,204,221;"         // blue
+        public var Warning = "fg253,202,78;"      // yellow
+        public var Error = "fg243,36,73;"         // red
     }
 
     var minLevelFilters = [MinLevelFilter]()
@@ -194,9 +196,10 @@ public class BaseDestination: Hashable, Equatable {
         var dateStr = ""
         var str = ""
         let levelStr = formattedLevel(level)
+        let formattedMsg = coloredMessage(msg, forLevel: level)
 
         dateStr = formattedDate(dateFormat)
-        str = formattedMessage(dateStr, levelString: levelStr, msg: msg, thread: thread, path: path,
+        str = formattedMessage(dateStr, levelString: levelStr, msg: formattedMsg, thread: thread, path: path,
             function: function, line: line, detailOutput: detailOutput)
         return str
     }
@@ -209,32 +212,62 @@ public class BaseDestination: Hashable, Equatable {
         return dateStr
     }
 
-    /// returns an optionally colored level noun (like INFO, etc.)
-    func formattedLevel(level: SwiftyBeaver.Level) -> String {
-        // optionally wrap the level string in color
+    /// returns the log message entirely colored
+    func coloredMessage(msg: String, forLevel level: SwiftyBeaver.Level) -> String {
+        if !(colored && coloredLines) {
+            return msg
+        }
+
+        let color = colorForLevel(level)
+        let coloredMsg = escape + color + msg + reset
+        return coloredMsg
+    }
+
+    /// returns color string for level
+    func colorForLevel(level: SwiftyBeaver.Level) -> String {
         var color = ""
-        var levelStr = ""
 
         switch level {
         case SwiftyBeaver.Level.Debug:
             color = levelColor.Debug
-            levelStr = levelString.Debug
 
         case SwiftyBeaver.Level.Info:
             color = levelColor.Info
-            levelStr = levelString.Info
 
         case SwiftyBeaver.Level.Warning:
             color = levelColor.Warning
-            levelStr = levelString.Warning
 
         case SwiftyBeaver.Level.Error:
             color = levelColor.Error
+
+        default:
+            color = levelColor.Verbose
+        }
+
+        return color
+    }
+
+    /// returns an optionally colored level noun (like INFO, etc.)
+    func formattedLevel(level: SwiftyBeaver.Level) -> String {
+        // optionally wrap the level string in color
+        let color = colorForLevel(level)
+        var levelStr = ""
+
+        switch level {
+        case SwiftyBeaver.Level.Debug:
+            levelStr = levelString.Debug
+
+        case SwiftyBeaver.Level.Info:
+            levelStr = levelString.Info
+
+        case SwiftyBeaver.Level.Warning:
+            levelStr = levelString.Warning
+
+        case SwiftyBeaver.Level.Error:
             levelStr = levelString.Error
 
         default:
             // Verbose is default
-            color = levelColor.Verbose
             levelStr = levelString.Verbose
         }